# server/pkg/api/rest/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collecting and Aggregating server module",
        "title": "CM-Honeybee Server REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/honeybee",
    "paths": {
        "/bench/{connId}": {
            "get": {
                "description": "Get the benchmark information of the connection information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] BenchmarkInfo"
                ],
                "summary": "Get Benchmark Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the connection info",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the benchmark.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedBenchmarkInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the benchmark.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bench/{connId}/run": {
            "get": {
                "description": "Run the benchmark information of the connection information. If no Benchmark Agent is present on the connected server, it will be automatically installed, and the benchmark will be executed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] BenchmarkInfo"
                ],
                "summary": "Run Benchmark Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the connection info",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get information of the benchmark.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.Benchmark"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get information of the benchmark.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bench/{connId}/stop": {
            "get": {
                "description": "Stop the benchmark",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] BenchmarkInfo"
                ],
                "summary": "Stop Benchmark",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the connection info",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Benchmark Stopped.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SimpleMsg"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to stop of the benchmark.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.Benchmark": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.BenchmarkData"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.BenchmarkData": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "elapsed": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "specid": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedBenchmarkInfo": {
            "type": "object",
            "properties": {
                "benchmark": {
                    "type": "string"
                },
                "connection_id": {
                    "type": "string"
                },
                "saved_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SimpleMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}